requirements.txt
aiogram>=3.0.0
playwright>=1.40.0
aiosqlite>=0.19.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
cryptography>=41.0.0
async-timeout>=4.0.2
pydantic>=2.5.0
redis>=5.0.0
orjson>=3.9.0
loguru>=0.7.2
